Saw strategy moved to deterministic.py. Updated some comments.
